Fixes for data output
Still need to fix some random misses and hton always returning 0

+++ src/common/data/bkv/states/gm_bkv_state_key.hpp
@@ -28,7 +28,7 @@ namespace game {
             static constexpr char DEFAULT_CHAR = -1;
             
             uint8_t key_[UINT8_MAX];
-            int64_t keyLen_ = 0;
+            int16_t keyLen_ = 0;
             char strChar_ = DEFAULT_CHAR;
             bool breakChar_ = false;
     };
